% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KORN_GRUBAURD.R, R/calculate_KG_CI.R
\name{calculate_KG_CI}
\alias{calculate_KG_CI}
\title{Calculate Korn and Graubard Confidence Intervals}
\usage{


calculate_KG_CI(
  df,
  prop_var = "proportion",
  se_var = "se",
  n_var = "counts",
  atlev1 = NULL,
  atlev2 = NULL,
  proportionTotal = 100,
  roundingIncrement = 0.1,
  df_adjust = TRUE
)
}
\arguments{
\item{df}{Input data frame}

\item{prop_var}{Name of proportion variable}

\item{se_var}{Name of standard error variable}

\item{n_var}{Name of unweighted count variable}

\item{atlev1}{Optional: number of strata variable}

\item{atlev2}{Optional: number of PSU variable}

\item{proportionTotal}{100 (default) or 1, depending on proportion scale}

\item{roundingIncrement}{Rounding increment for display}

\item{df_adjust}{Whether to adjust for degrees of freedom (default: TRUE)}
}
\value{
Data frame with calculated CIs and flags
}
\description{
Calculate Korn and Graubard Confidence Intervals

}
